partial refactor: move directive functionality into metaclass that is intended to be shared between system and experiment
