+++ lib/benchpark/system.py
@@ -7,6 +7,7 @@ import pathlib
 import sys
 
 import benchpark.paths
+from benchpark.directives import SpecTemplate
 import benchpark.repo
 from benchpark.runtime import RuntimeResources
 
@@ -66,115 +67,6 @@ compilers_schema = load_schema(
 _repo_path = benchpark.repo.paths[benchpark.repo.ObjectTypes.systems]
 
 
-# Duplicated from experiment.py
-class classproperty:
-    """Non-data descriptor to evaluate a class-level property. The function that performs
-    the evaluation is injected at creation time and take an instance (could be None) and
-    an owner (i.e. the class that originated the instance)
-    """
-
-    def __init__(self, callback):
-        self.callback = callback
-
-    def __get__(self, instance, owner):
-        return self.callback(owner)
-
-
-class SystemMeta(ramble.language.shared_language.SharedMeta):
-    _directive_names = set()
-    _directives_to_be_executed = []
-
-    # Hack to be able to use SharedMeta outside of Ramble
-    # will ask Ramble to implement fix on their end and then we can remove this
-    def __init__(self, *args, **kwargs):
-        with benchpark.repo.override_ramble_hardcoded_globals():
-            super(SystemMeta, self).__init__(*args, **kwargs)
-
-
-system_directive = SystemMeta.directive
-
-
-@system_directive("variants")
-def variant(
-    name: str,
-    default: Optional[Any] = None,
-    description: str = "",
-    values: Optional[Union[collections.abc.Sequence, Callable[[Any], bool]]] = None,
-    multi: Optional[bool] = None,
-    validator: Optional[Callable[[str, str, Tuple[Any, ...]], None]] = None,
-    when: Optional[Union[str, bool]] = None,
-    sticky: bool = False,
-):
-    def format_error(msg, pkg):
-        msg += " @*r{{[{0}, variant '{1}']}}"
-        return msg.format(pkg.name, name)
-
-    def _always_true(_x):
-        return True
-
-    # Ensure we have a sequence of allowed variant values, or a
-    # predicate for it.
-    if values is None:
-        if str(default).upper() in ("TRUE", "FALSE"):
-            values = (True, False)
-        else:
-            values = _always_true
-
-    # The object defining variant values might supply its own defaults for
-    # all the other arguments. Ensure we have no conflicting definitions
-    # in place.
-    for argument in ("default", "multi", "validator"):
-        # TODO: we can consider treating 'default' differently from other
-        # TODO: attributes and let a packager decide whether to use the fluent
-        # TODO: interface or the directive argument
-        if hasattr(values, argument) and locals()[argument] is not None:
-
-            def _raise_argument_error(pkg):
-                msg = (
-                    "Remove specification of {0} argument: it is handled "
-                    "by an attribute of the 'values' argument"
-                )
-                raise DirectiveError(format_error(msg.format(argument), pkg))
-
-            return _raise_argument_error
-
-    # Allow for the object defining the allowed values to supply its own
-    # default value and group validator, say if it supports multiple values.
-    default = getattr(values, "default", default)
-    validator = getattr(values, "validator", validator)
-    multi = getattr(values, "multi", bool(multi))
-
-    # Here we sanitize against a default value being either None
-    # or the empty string, as the former indicates that a default
-    # was not set while the latter will make the variant unparsable
-    # from the command line
-    if default is None or default == "":
-
-        def _raise_default_not_set(pkg):
-            if default is None:
-                msg = "either a default was not explicitly set, or 'None' was used"
-            elif default == "":
-                msg = "the default cannot be an empty string"
-            raise DirectiveError(format_error(msg, pkg))
-
-        return _raise_default_not_set
-
-    description = str(description).strip()
-
-    def _execute_variant(pkg):
-        if not re.match(benchpark.spec.IDENTIFIER, name):
-            directive = "variant"
-            msg = "Invalid variant name in {0}: '{1}'"
-            raise DirectiveError(directive, msg.format(pkg.name, name))
-
-        pkg.variants[name] = benchpark.variant.Variant(
-            name, default, description, values, multi, validator, sticky
-        )
-
-    return _execute_variant
-# end duplication from experiment.py
-
-
 def _hash_id(content_list):
     sha256_hash = hashlib.sha256()
     for x in content_list:
@@ -182,7 +74,7 @@ def _hash_id(content_list):
     return sha256_hash.hexdigest()
 
 
-class System(metaclass=SystemMeta):
+class System(SpecTemplate):
     variants: Dict[
         str,
         Tuple["benchpark.variant.Variant", "benchpark.spec.ConcreteSystemSpec"],
@@ -191,43 +83,6 @@ class System(metaclass=SystemMeta):
     def __init__(self, spec):
         self.spec: "benchpark.spec.ConcreteSystemSpec" = spec
         super().__init__()
-
-    @classproperty
-    def package_dir(cls):
-        return os.path.abspath(os.path.dirname(cls.module.__file__))
-
-    @classproperty
-    def module(cls):
-        return __import__(cls.__module__, fromlist=[cls.__name__])
-
-    @classproperty
-    def namespace(cls):
-        parts = cls.__module__.split(".")
-        return ".".join(parts[2:-1])
-
-    @classproperty
-    def fullname(cls):
-        return f"{cls.namespace}.{cls.name}"
-
-    @classproperty
-    def fullnames(cls):
-        fullnames = []
-        for cls in inspect.getmro(cls):
-            namespace = getattr(cls, "namespace", None)
-            if namespace:
-                fullnames.append(f"{namespace}.{cls.name}")
-            if namespace == "builtin":
-                # builtin packages cannot inherit from other repos
-                break
-        return fullnames
-
-    @classproperty
-    def name(cls):
-        if cls._name is None:
-            cls._name = cls.module.__name__
-            if "." in cls._name:
-                cls._name = cls._name[cls._name.rindex(".") + 1 :]
-        return cls._name
 
     def initialize(self):
         self.external_resources = None
